requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.2
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==21.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
windows-curses==2.3.0
